fixed: more gcc warnings, in DVDDemuxFFmpeg delete[] was called on a void*.  According to gcc, this is undefined; so I commented it out.  If this is incorrect, please revert this change.
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@11423 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/lib/libPython/xbmcmodule/GUIPythonWindowXML.cpp
@@ -98,7 +98,7 @@ bool CGUIPythonWindowXML::OnMessage(CGUIMessage& message)
 
     case GUI_MSG_SETFOCUS:
     {
-      if (m_viewControl.HasControl(message.GetControlId()) && m_viewControl.GetCurrentControl() != message.GetControlId())
+      if (m_viewControl.HasControl(message.GetControlId()) && m_viewControl.GetCurrentControl() != (int)message.GetControlId())
       {
         m_viewControl.SetFocused();
         return true;
@@ -150,7 +150,7 @@ bool CGUIPythonWindowXML::OnMessage(CGUIMessage& message)
         return true;
       }
 
-      if(pCallbackWindow && iControl && iControl != this->GetID()) // pCallbackWindow &&  != this->GetID())
+      if(pCallbackWindow && iControl && iControl != (int)this->GetID()) // pCallbackWindow &&  != this->GetID())
       {
         CGUIControl* controlClicked = (CGUIControl*)this->GetControl(iControl);
